cleaned up some more

+++ app/src/main/java/com/ihongqiqu/util/BitmapUtil.java
@@ -1430,9 +1430,9 @@ public final class BitmapUtil {
                         pixG = Color.green(pixColor);
                         pixB = Color.blue(pixColor);
 
-                        newR = newR + (int) (pixR * gauss[idx]);
-                        newG = newG + (int) (pixG * gauss[idx]);
-                        newB = newB + (int) (pixB * gauss[idx]);
+                        newR = newR + pixR * gauss[idx];
+                        newG = newG + pixG * gauss[idx];
+                        newB = newB + pixB * gauss[idx];
                         idx++;
                     }
                 }
